requirements.txt
openai
langchain
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
